+++ Bitkit/MainNavView.swift
@@ -479,25 +479,6 @@ struct MainNavView: View {
 
                 // Widgets
                 case .widgetsIntro: WidgetsIntroView()
-                case .widgetsList: WidgetsListView()
-                case let .widgetDetail(widgetType):
-                    switch widgetType {
-                    case .price:
-                        PriceWidgetPreviewView()
-                    case .news:
-                        NewsWidgetPreviewView()
-                    case .blocks:
-                        BlocksWidgetPreviewView()
-                    case .facts:
-                        FactsWidgetPreviewView()
-                    case .weather:
-                        WeatherWidgetPreviewView()
-                    case .calculator:
-                        CalculatorWidgetPreviewView()
-                    default:
-                        WidgetDetailView(id: widgetType)
-                    }
-                case let .widgetEdit(widgetType): WidgetEditView(id: widgetType)
 
                 // Settings
                 case .settings: MainSettingsScreen()
