+++ Bitkit/Views/Contacts/EditContactView.swift
@@ -62,7 +62,6 @@ struct EditContactView: View {
 
     // MARK: - Avatar
 
-    @ViewBuilder
     private var avatarSection: some View {
         PhotosPicker(selection: $selectedPhotoItem, matching: .images) {
             Group {
